perf(next): split diagnostics by runtime
