+++ .changeset/split-next-diagnostics.md
@@ -1,0 +1,5 @@
+---
+'gt-next': patch
+---
+
+Split browser, request, middleware, and loader diagnostics into runtime-specific modules.
