+++ packages/next/src/errors/middleware.ts
@@ -1,0 +1,9 @@
+import { getLocaleProperties } from '@generaltranslation/format';
+
+export const createUnsupportedLocalesWarning = (locales: string[]) =>
+  `gt-next: The following locales are currently unsupported by our service: ${locales
+    .map((locale) => {
+      const { name } = getLocaleProperties(locale);
+      return `${locale} (${name})`;
+    })
+    .join(', ')}`;
