+++ packages/next/src/errors/pathRegex.ts
@@ -1,0 +1,15 @@
+import {
+  createGtNextDiagnostic,
+  formatDiagnosticErrorDetails,
+} from './diagnostics';
+
+export const createInvalidPathRegexError = (
+  pathRegex: string,
+  error: unknown
+) =>
+  createGtNextDiagnostic({
+    severity: 'Error',
+    whatHappened: `pathRegex "${pathRegex}" is not a valid regular expression`,
+    fix: 'Pass a valid JavaScript regular expression string to withGTConfig()',
+    details: formatDiagnosticErrorDetails(error),
+  });
